# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartDev LMS API Documentation",
        "description": "Comprehensive API documentation for SmartDev Academic Learning Management System. This API provides endpoints for managing courses, students, instructors, assignments, submissions, grades, attendance, and more.",
        "contact": {
            "name": "SmartDev LMS Support",
            "email": "admin@smartdevlms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://portohansgunawan.my.id",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/enrollments": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get all enrollments",
                "description": "Retrieve a list of all enrollments with student and course details",
                "operationId": "5b0e4ac73de1605b97b76c9b18818440",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "student_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "course_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "enrolled_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Enroll a student in a course",
                "description": "Create a new enrollment for a student in a specific course",
                "operationId": "1a36b2c6aa9b4a8177e474657f262dcd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "course_id"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enrollment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "student_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "course_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "enrolled_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/enrollments/{id}": {
            "get": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Get enrollment by ID",
                "description": "Retrieve a specific enrollment with student and course details",
                "operationId": "2fdac9d470f0c0e43feecabdaddb3283",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "student_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "course_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "enrolled_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Enrollment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Update enrollment (not typically supported)",
                "description": "Enrollments are generally not updated, only created or deleted",
                "operationId": "7d70b01af9957d8164eb007f78b2ee1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Method not allowed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Enrollment updates are not typically supported."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Delete enrollment",
                "description": "Remove a student's enrollment from a course",
                "operationId": "c96fa6289f1ec321b94d251a6463604b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enrollment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Enrollment deleted successfully"
                    },
                    "404": {
                        "description": "Enrollment not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://portohansgunawan.my.id/oauth/authorize",
                        "tokenUrl": "http://portohansgunawan.my.id/oauth/token",
                        "refreshUrl": "http://portohansgunawan.my.id/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}